+++ src/main/java/com/ood/Model/Currency/JPY_Currency.java
@@ -12,7 +12,7 @@ public class JPY_Currency extends AbsCurrency{
         this.amount-=valueOf(otherAmount,type);
     }
 
-    public static float valueOf(float otherAmount, CurrencyEnum type) {
+    public static double valueOf(double otherAmount, CurrencyEnum type) {
         if(type.equals(CurrencyEnum.USD))
         {
             return otherAmount/Constants.EXCHANGE_RATE_JPY_DIV_USD;
